Add vcxproj test projects for 4 C++ test suites
Create NativeUnitTestProject DLL projects for:
- Runner HotkeyConflictTests (src/runner/UnitTests/UnitTests-Runner.vcxproj)
- CursorWrap TopologyTests (src/modules/MouseUtils/CursorWrap/UnitTests/UnitTests-CursorWrap.vcxproj)
- MouseHighlighter HighlighterTests (src/modules/MouseUtils/MouseHighlighter/UnitTests/UnitTests-MouseHighlighter.vcxproj)
- MousePointerCrosshairs CrosshairsTests (src/modules/MouseUtils/MousePointerCrosshairs/UnitTests/UnitTests-Crosshairs.vcxproj)

Each project follows the UnitTests-CommonLib pattern with:
- DynamicLibrary configuration for MSTest native test DLLs
- NativeUnitTestProject subtype
- CppUnitTestFramework reference via UnitTest lib paths
- Precompiled headers (pch.h/pch.cpp)
- Proper include directories for module headers and common libs

Move test .cpp files into UnitTests/ subdirectories, add all
projects to PowerToys.slnx, and fix code analysis warnings
(C26451 overflow, C26497 constexpr, C4189 unused variable).

+++ src/modules/MouseUtils/MouseHighlighter/UnitTests/HighlighterTests.cpp
@@ -17,7 +17,7 @@
 #include "CppUnitTest.h"
 #pragma warning(pop)
 
-#include "MouseHighlighter.h"
+#include "../MouseHighlighter.h"
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
@@ -93,7 +93,7 @@ namespace MouseHighlighterUnitTests
             //   clamped to [0.0, 1.0]
             // At elapsed == delay + duration: 1.0 - duration/duration = 0.0
             int elapsed = settings.fadeDelayMs + settings.fadeDurationMs;
-            double opacity = 1.0 - static_cast<double>(elapsed - settings.fadeDelayMs) / settings.fadeDurationMs;
+            double opacity = 1.0 - (static_cast<double>(elapsed) - static_cast<double>(settings.fadeDelayMs)) / static_cast<double>(settings.fadeDurationMs);
             Assert::AreEqual(0.0, opacity, 1e-10,
                              L"Opacity should be 0.0 at delay+duration");
         }
